% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biogrid.R
\name{getBiogridSymbols}
\alias{getBiogridSymbols}
\title{Extracts unique Biogrid Symbols}
\usage{
getBiogridSymbols(biogrid)
}
\arguments{
\item{biogrid}{A data frame with the BioGRID data}
}
\value{
A data frame with one column with the unique BioGRID symbols.
}
\description{
\code{getBiogridSymbols} returns a list of unique BioGRID symbols
}
\details{
The symbols are extracted from the columns: \code{Official.Symbol.Interactor.A},
\code{Official.Symbol.Interactor.B}, \code{Synonyms.Interactor.A}, and
\code{Synonyms.Interactor.B}.
}
\examples{
biogrid_symbols <- getBiogridSymbols(biogrid)
}
