% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPIfinder.R
\name{getCompleteSymbols}
\alias{getCompleteSymbols}
\title{Gene name finder}
\usage{
getCompleteSymbols(data, species)
}
\arguments{
\item{data}{a data frame that contains contains gene symbols}
}
\value{
The function \code{getCompleteSymbols} returns a data frame with a single column that lists all unique symbols, symbol synonyms, and names, and interpor protein, complex, and domain names.
}
\description{
\code{getCompleteSymbols} extracts all symbols from a data frame and finds their gene names, synonymns, and interpor protein, complex, and domain names. The function returns a data frame with a single column that lists all unique symbols, symbol synonyms, and names, and interpor protein, complex, and domain names. The species must be specified by the user as a string (e.g., "Human").
}
\details{
This function requires the \code{mygene} library and the \code{getPatternList} function.
}
\examples{
## get a single list of all symbols and names from a data frame
 symbols <- getCompleteSymbols(data)
}
