% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPIfinder.R
\name{getInteractionMatches}
\alias{getInteractionMatches}
\title{Identifies user-defined keywords and sentences containing these keywords}
\usage{
getInteractionMatches(data, keywords)
}
\arguments{
\item{data}{a character vector of sentences separated by '|'.}

\item{keywords}{a list of keywords to identify protein-protein interactions.}
}
\value{
The function \code{getInteractionMatches} returns a nested list with two lists: sentences containing keywords, keywords. If matched sentences do not contain any of the provided keywords the cell will contain "No keywords found in sentence" and or "No keywords". The nested list will contain the following sub-lists: sents-Sentences from abstract containing a matched symbol and one or more of the keywords, separated by '|'; kws-Matched keywords from int_sentences, separated by '|'.
}
\description{
\code{getInteractionMatches} searches among each sentence that contains two proteins using a user provided list of keywords for matches. If a match is found, the sentences containing the keywords and the matched keywords are returned as a nested list.
}
\details{
This function requires the \code{quanteda} library.
}
\examples{
## identifiying sentences contiaing specific keywords
interaction_info <- getInteractionMatches(data, keywords)
}
