% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPIfinder.R
\name{getPPIs}
\alias{getPPIs}
\title{Find Protein-Protein Interactions}
\usage{
getPPIs(data, regex, getInteractionMatches, keywords)
}
\arguments{
\item{data}{A data frame containing PubMed Ids, gene A and B symbols, synonyms list (each symbol separated by '|'), gene A and B names, and article title and abstract.}

\item{regex}{a large data frame of symbols and regular expression patterns.}

\item{getInteractionMatches}{- see specific function documentation \code{nlpUtilityBelt::getInteractionMatches}.}

\item{keywords}{a list of keywords to identify PPIs.}
}
\value{
The function "getPPIs" returns a data frame with seven columns: pmid, article_title, article_abstract, matched_symbols, matched_setences, int_sentences, int_keywords, and match. If matched sentences do not contain any of the provided keywords the cell will contain "No keywords found in sentence" and or "No keywords". The data frame will contain the following columns: pmid-PubMed IDs; article_title-PubMed article title; article_abstract-PubMed article abstract sentences separated by '|'; matched_symbols-A list of identified protein symbols separated by '|'; matched_sentences-Sentences from abstract containing a matched symbol, separated by '|'; int_sentences-Sentences from abstract containing a matched symbol and one or more of the keywords, separated by '|'; int_keywords-Matched keywords from int_sentences, separated by '|'; match-A '0' or '1' to indicate articles where gene A and B where identified
}
\description{
\code{getPPIs} identifies proteins in PubMed titles and abstracts and if a match is found returns matching information as data frame. The function searches among each sentence that contains two proteins using a user provided list of keywords for matches. If a match is found, the sentences containing the keywords and the matched keywords are appended to data frame.
}
\details{
This function requires the packages \code{stringr}. It also uses the \code{getInteractionMacthes} and \code{extractPOS} functions.
}
\examples{
## interaction keywords
keywords <- c("bind", 
             "interact",
             "associate",
             "regulation",
             "bound",
             "localize",
             "stimulation",
             "regulate",
             "effect",
             "target",
             "component",
             "member",
             "mediate")

## loop over abstracts - get subset of the data for testing
PPI_results <- getPPIs(merged_biogrid_pubmed_results, patterns, getInteractionMatches, keywords)

## write out sentences
write.table(PPI_results, "PPI_BIOGRID_results.txt", quote  = FALSE, sep = '\\t', col.names = TRUE, row.names = FALSE)
}
