% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{getPubmedAbstracts}
\alias{getPubmedAbstracts}
\title{Extracts abstracts and titles from PUBMED from a list of selected PMIDs}
\usage{
getPubmedAbstracts(pmids)
}
\arguments{
\item{pmids}{A list of Pubmed IDs}
}
\value{
A dataframe with three columns: PMID, Title, and Abstract
}
\description{
\code{getPubmedAbstracts} extracts a list of abstracts from Pubmed for specific PMIDs.
}
\details{
This function requires the package "RISmed".
}
\examples{
## getting Pubmed abstracts
pmids <- c("7545544", "9342365")
abstracts <- getPubmedAbstracts(pmids)
abstracts[1,]
}
