% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubtator.R
\name{getPubtatorMatches}
\alias{getPubtatorMatches}
\title{Find Pubtator information on Pubmed abstracts}
\usage{
getPubtatorMatches(abstracts, info = c("Genes", "Diseases", "Mutations",
  "Chemicals", "Species"))
}
\arguments{
\item{abstracts}{A data frame with a list of Pubmed abstracts, including:
PMID, Title and Abstract}

\item{info}{Check what PubTator information to use for list of symbols,
including genes, diseases, mutations, chemicals and species}
}
\value{
The function "getPubtatorMatches" returns a data frame with four columns:
PMID, symbols, title, and sentences. If the regex encounters an error, it will
return the string "*** REGEX ERROR ***".
The column "symbols" represent the Pubtator info found in the astract or title.
The symbols are separated by a pipe "|".
The column "title" will show the title only if there are at least 2 symbols in it.
The colum "sentences" will show all sentences from the abstract that have at
least two occurrences of any symbol listed in the column "symbols". The sentences are separated by a pipe "|".
}
\description{
\code{getPubtatorMatches} finds selected Pubtator information
that occurs at least twice in the title and in the abstract.
}
\details{
This function requires the packages "pubmed.mineR" and "plyr".
}
\examples{
## getting Pubtator matches from abstracts extracted from Pubmed
pubtator <- getPubtatorMatches(abstracts,
                               info = c("genes"))
pubtator[1,]
}
