% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPIfinder.R
\name{matchFinder}
\alias{matchFinder}
\title{Identifies symbols and sentences containing the symbols in a match matrix}
\usage{
matchFinder(sentence_matches, regex, sentences)
}
\arguments{
\item{sentence_matches}{A matrix of matches, where values in the matrix are generated by the results of running grepl against all potential symbol matches and the title and abstract of a journal article.}

\item{regex}{a large data frame of symbols and regular expression patterns.}

\item{sentences}{a character vetor of sentences, the title and abstract.}
}
\value{
The function \code{matchFinder} The function returns a nested list where the first list contains matched sentences and the second contains matched symbols.
}
\description{
\code{matchFinder} searches through each row (sentence) and column (symbol) of a matrix of matches. Values in the matrix are generated by the results of running grepl against all potential symbol matches and the title and abstract of a journal article. Additional filtering is performed to remove columns where the symbol is not in a master symbol list. The function returns a nested list where the first list contains matched sentences and the second contains matched symbols.
}
\examples{
## identifiying sentences contiaing specific keywords
matches <- matchFinder(sentence_matches, regex, sentences)
}
