% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{matchTokens}
\alias{matchTokens}
\title{Match tokens}
\usage{
matchTokens(patternList, selTokens, fixed = FALSE, window = 500)
}
\arguments{
\item{patternList}{}

\item{selTokens}{}

\item{fixed}{}
}
\examples{

}
