% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{parse2graph}
\alias{parse2graph}
\title{Make a graph from Tree_parse result}
\usage{
parse2graph(ptext, leaf.color = "chartreuse4", label.color = "blue4",
  title = NULL, cex.main = 0.9, ...)
}
\arguments{
\item{ptext}{}

\item{leaf.color}{}

\item{label.color}{}

\item{title}{}

\item{cex.main}{}
}
\description{
Source: \url{http://stackoverflow.com/questions/33473107/visualize-parse-tree-structure/33536291}
}
\examples{

}
