% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{setPatternRange}
\alias{setPatternRange}
\title{Set range for similar regular expressions}
\usage{
setPatternRange(pl, colname = "symbols")
}
\arguments{
\item{pl}{A list of patterns}

\item{colname}{The name of the column containing the symbols}
}
\value{
a dataframe with patterns converted to restrict format.
The restricted option combines the number of patterns using the min and max
number of occurrencens. For example: [a-zA-Z]{2-3}
}
\description{
\code{setPatternRange} returns a data frame with the list of
restriced patterns.
}
\examples{
## getting patterns with ranges
patternRanges <- getPatternsRange(patternsList)
patternRanges[1,]
}
